Cleanup imports and unused code
