+++ archinstall/lib/hardware.py
@@ -1,8 +1,8 @@
 import os
 import logging
-from functools import partial
+from functools import cached_property
 from pathlib import Path
-from typing import Iterator, Optional, Dict
+from typing import Optional, Dict
 
 from .general import SysCommand
 from .networking import list_interfaces, enrich_iface_types
@@ -49,137 +49,125 @@ AVAILABLE_GFX_DRIVERS = {
 }
 
 
+class _SysInfo:
+	def __init__(self):
+		pass
 
-def cpu_info() -> Iterator[dict[str, str]]:
-	"""
-	Yields information about the CPUs of the system
-	"""
-	cpu_info_path = Path("/proc/cpuinfo")
-	cpu: Dict[str, str] = {}
+	@cached_property
+	def cpu_info(self) -> Dict[str, str]:
+		"""
+		Returns system cpu information
+		"""
+		cpu_info_path = Path("/proc/cpuinfo")
+		cpu: Dict[str, str] = {}
 
-	with cpu_info_path.open() as file:
-		for line in file:
-			if not (line := line.strip()):
-				yield cpu
-				cpu = {}
-				continue
+		with cpu_info_path.open() as file:
+			for line in file:
+				if line := line.strip():
+					key, value = line.split(":", maxsplit=1)
+					cpu[key.strip()] = value.strip()
 
-			key, value = line.split(":", maxsplit=1)
-			cpu[key.strip()] = value.strip()
+		return cpu
 
+	@cached_property
+	def mem_info(self) -> Dict[str, int]:
+		"""
+		Returns system memory information
+		"""
+		mem_info_path = Path("/proc/meminfo")
+		mem_info: Dict[str, int] = {}
 
-def all_meminfo() -> Dict[str, int]:
-	"""
-	Returns a dict with memory info if called with no args
-	or the value of the given key of said dict.
-	"""
-	mem_info_path = Path("/proc/meminfo")
-	mem_info: Dict[str, int] = {}
+		with mem_info_path.open() as file:
+			for line in file:
+				key, value = line.strip().split(':')
+				num = value.split()[0]
+				mem_info[key] = int(num)
 
-	with mem_info_path.open() as file:
-		for line in file:
-			key, value = line.strip().split(':')
-			num = value.split()[0]
-			mem_info[key] = int(num)
+		return mem_info
 
-	return mem_info
-
-
-def _meminfo_for_key(key: str) -> int:
-	info = all_meminfo()
-	return info[key]
-
-
-def has_wifi() -> bool:
-	ifaces = list(list_interfaces().values())
-	return 'WIRELESS' in enrich_iface_types(ifaces).values()
+	def mem_info_by_key(self, key: str) -> int:
+		return self.mem_info[key]
 
 
-def has_cpu_vendor(vendor_id: str) -> bool:
-	return any(cpu.get("vendor_id") == vendor_id for cpu in cpu_info())
+_sys_info = _SysInfo()
 
 
-has_amd_cpu = partial(has_cpu_vendor, "AuthenticAMD")
+class SysInfo:
+	@staticmethod
+	def has_wifi() -> bool:
+		ifaces = list(list_interfaces().values())
+		return 'WIRELESS' in enrich_iface_types(ifaces).values()
 
+	@staticmethod
+	def has_uefi() -> bool:
+		return os.path.isdir('/sys/firmware/efi')
 
-has_intel_cpu = partial(has_cpu_vendor, "GenuineIntel")
+	@staticmethod
+	def _graphics_devices() -> Dict[str, str]:
+		cards: Dict[str, str] = {}
+		for line in SysCommand("lspci"):
+			if b' VGA ' in line or b' 3D ' in line:
+				_, identifier = line.split(b': ', 1)
+				cards[identifier.strip().decode('UTF-8')] = str(line)
+		return cards
 
+	@staticmethod
+	def has_nvidia_graphics() -> bool:
+		return any('nvidia' in x.lower() for x in SysInfo._graphics_devices())
 
-def has_uefi() -> bool:
-	return os.path.isdir('/sys/firmware/efi')
+	@staticmethod
+	def has_amd_graphics() -> bool:
+		return any('amd' in x.lower() for x in SysInfo._graphics_devices())
 
+	@staticmethod
+	def has_intel_graphics() -> bool:
+		return any('intel' in x.lower() for x in SysInfo._graphics_devices())
 
-def graphics_devices() -> dict:
-	cards = {}
-	for line in SysCommand("lspci"):
-		if b' VGA ' in line or b' 3D ' in line:
-			_, identifier = line.split(b': ', 1)
-			cards[identifier.strip().decode('UTF-8')] = line
-	return cards
+	@staticmethod
+	def cpu_vendor() -> Optional[str]:
+		return _sys_info.cpu_info.get('vendor_id', None)
 
+	@staticmethod
+	def cpu_model() -> Optional[str]:
+		return _sys_info.cpu_info.get('model name', None)
 
-def has_nvidia_graphics() -> bool:
-	return any('nvidia' in x.lower() for x in graphics_devices())
+	@staticmethod
+	def sys_vendor() -> str:
+		with open(f"/sys/devices/virtual/dmi/id/sys_vendor") as vendor:
+			return vendor.read().strip()
 
+	@staticmethod
+	def product_name() -> str:
+		with open(f"/sys/devices/virtual/dmi/id/product_name") as product:
+			return product.read().strip()
 
-def has_amd_graphics() -> bool:
-	return any('amd' in x.lower() for x in graphics_devices())
+	@staticmethod
+	def mem_available() -> int:
+		return _sys_info.mem_info_by_key('MemAvailable')
 
+	@staticmethod
+	def mem_free() -> int:
+		return _sys_info.mem_info_by_key('MemFree')
 
-def has_intel_graphics() -> bool:
-	return any('intel' in x.lower() for x in graphics_devices())
+	@staticmethod
+	def mem_total() -> int:
+		return _sys_info.mem_info_by_key('MemTotal')
 
+	@staticmethod
+	def virtualization() -> Optional[str]:
+		try:
+			return str(SysCommand("systemd-detect-virt")).strip('\r\n')
+		except SysCallError as error:
+			log(f"Could not detect virtual system: {error}", level=logging.DEBUG)
 
-def cpu_vendor() -> Optional[str]:
-	for cpu in cpu_info():
-		return cpu.get("vendor_id")
+		return None
 
-	return None
+	@staticmethod
+	def is_vm() -> bool:
+		try:
+			result = SysCommand("systemd-detect-virt")
+			return b"none" not in b"".join(result).lower()
+		except SysCallError as error:
+			log(f"System is not running in a VM: {error}", level=logging.DEBUG)
 
-
-def cpu_model() -> Optional[str]:
-	for cpu in cpu_info():
-		return cpu.get("model name")
-
-	return None
-
-
-def sys_vendor() -> Optional[str]:
-	with open(f"/sys/devices/virtual/dmi/id/sys_vendor") as vendor:
-		return vendor.read().strip()
-
-
-def product_name() -> Optional[str]:
-	with open(f"/sys/devices/virtual/dmi/id/product_name") as product:
-		return product.read().strip()
-
-
-def mem_available() -> Optional[int]:
-	return _meminfo_for_key('MemAvailable')
-
-
-def mem_free() -> Optional[int]:
-	return _meminfo_for_key('MemFree')
-
-
-def mem_total() -> Optional[int]:
-	return _meminfo_for_key('MemTotal')
-
-
-def virtualization() -> Optional[str]:
-	try:
-		return str(SysCommand("systemd-detect-virt")).strip('\r\n')
-	except SysCallError as error:
-		log(f"Could not detect virtual system: {error}", level=logging.DEBUG)
-
-	return None
-
-
-def is_vm() -> bool:
-	try:
-		result = SysCommand("systemd-detect-virt")
-		return b"none" not in b"".join(result).lower()
-	except SysCallError as error:
-		log(f"System is not running in a VM: {error}", level=logging.DEBUG)
-
-	return False
+		return False
