+++ archinstall/lib/interactions/__init__.py
@@ -1,0 +1,20 @@
+from .locale_conf import select_locale_lang, select_locale_enc
+from .manage_users_conf import UserList, ask_for_additional_users
+from .network_conf import ManualNetworkConfig, ask_to_configure_network
+from .utils import get_password
+
+from .disk_conf import (
+	select_devices, select_disk_config, get_default_partition_layout,
+	select_main_filesystem_format, suggest_single_disk_layout,
+	suggest_multi_disk_layout
+)
+
+from .general_conf import (
+	ask_ntp, ask_hostname, ask_for_a_timezone, ask_for_audio_selection, select_language,
+	select_mirror_regions, select_archinstall_language, ask_additional_packages_to_install,
+	add_number_of_parrallel_downloads, select_additional_repositories
+)
+
+from .system_conf import (
+	select_kernel, ask_for_bootloader, select_driver, ask_for_swap
+)
