+++ archinstall/lib/interactions/general_conf.py
@@ -4,7 +4,7 @@ import logging
 import pathlib
 from typing import List, Any, Optional, Dict, TYPE_CHECKING
 
-from ..locale_helpers import list_keyboard_languages, list_timezones
+from ..locale import list_keyboard_languages, list_timezones
 from ..menu import MenuSelectionType, Menu, TextInput
 from ..mirrors import list_mirrors
 from ..output import log
