+++ archinstall/lib/interactions/utils.py
