+++ archinstall/scripts/swiss.py
@@ -5,14 +5,16 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Any, Dict
 
 import archinstall
+from archinstall import SysInfo
 from archinstall.lib.mirrors import use_mirrors
-from archinstall import models
-from archinstall import disk
+from archinstall.lib import models
+from archinstall.lib import disk
+from archinstall.lib.networking import check_mirror_reachable
 from archinstall.lib.profile.profiles_handler import profile_handler
-from archinstall import menu
+from archinstall.lib import menu
 from archinstall.lib.global_menu import GlobalMenu
 from archinstall.lib.output import log
-from archinstall import Installer
+from archinstall.lib.installer import Installer
 from archinstall.lib.configuration import ConfigurationOutput
 from archinstall.default_profiles.applications.pipewire import PipewireProfile
 
@@ -23,11 +25,6 @@ if TYPE_CHECKING:
 if archinstall.arguments.get('help'):
 	print("See `man archinstall` for help.")
 	exit(0)
-
-
-if os.getuid() != 0:
-	print("Archinstall requires root privileges to run. See --help for more.")
-	exit(1)
 
 
 class ExecutionMode(Enum):
@@ -219,7 +216,7 @@ def perform_installation(mountpoint: Path, exec_mode: ExecutionMode):
 			if archinstall.arguments.get('swap'):
 				installation.setup_swap('zram')
 
-			if archinstall.arguments.get("bootloader") == models.Bootloader.Grub and archinstall.has_uefi():
+			if archinstall.arguments.get("bootloader") == models.Bootloader.Grub and SysInfo.has_uefi():
 				installation.add_additional_packages("grub")
 
 			installation.add_bootloader(archinstall.arguments["bootloader"])
@@ -300,18 +297,7 @@ def perform_installation(mountpoint: Path, exec_mode: ExecutionMode):
 		archinstall.log(f"Disk states after installing: {disk.disk_layouts()}", level=logging.DEBUG)
 
 
-# Log various information about hardware before starting the installation. This might assist in troubleshooting
-archinstall.log(f"Hardware model detected: {archinstall.sys_vendor()} {archinstall.product_name()}; UEFI mode: {archinstall.has_uefi()}", level=logging.DEBUG)
-archinstall.log(f"Processor model detected: {archinstall.cpu_model()}", level=logging.DEBUG)
-archinstall.log(f"Memory statistics: {archinstall.mem_available()} available out of {archinstall.mem_total()} total installed", level=logging.DEBUG)
-archinstall.log(f"Virtualization detected: {archinstall.virtualization()}; is VM: {archinstall.is_vm()}", level=logging.DEBUG)
-archinstall.log(f"Graphics devices detected: {archinstall.graphics_devices().keys()}", level=logging.DEBUG)
-
-# For support reasons, we'll log the disk layout pre installation to match against post-installation layout
-archinstall.log(f"Disk states before installing: {disk.disk_layouts()}", level=logging.DEBUG)
-
-
-if not archinstall.check_mirror_reachable():
+if not check_mirror_reachable():
 	log_file = os.path.join(archinstall.storage.get('LOG_PATH', None), archinstall.storage.get('LOG_FILE', None))
 	archinstall.log(f"Arch Linux mirrors are not reachable. Please check your internet connection and the log file '{log_file}'.", level=logging.INFO, fg="red")
 	exit(1)
