+++ examples/full_automated_installation.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 
-from archinstall import Installer, ProfileConfiguration, profile_handler
+from archinstall import Installer
+from archinstall import profile
 from archinstall.default_profiles.minimal import MinimalProfile
 from archinstall import disk
 from archinstall.lib.models import User
@@ -88,8 +89,8 @@ with Installer(
 
 # Optionally, install a profile of choice.
 # In this case, we install a minimal profile that is empty
-profile_config = ProfileConfiguration(MinimalProfile())
-profile_handler.install_profile_config(installation, profile_config)
+profile_config = profile.ProfileConfiguration(MinimalProfile())
+profile.profile_handler.install_profile_config(installation, profile_config)
 
 user = User('archinstall', 'password', True)
 installation.create_users(user)
