+++ examples/mac_address_installation.py
@@ -1,13 +1,13 @@
 import time
 
 import archinstall
-from archinstall.lib.profile.profiles_handler import profile_handler
+from archinstall import profile
 
-for profile in profile_handler.get_mac_addr_profiles():
+for _profile in profile.profile_handler.get_mac_addr_profiles():
 	# Tailored means it's a match for this machine
 	# based on it's MAC address (or some other criteria
 	# that fits the requirements for this machine specifically).
-	archinstall.log(f'Found a tailored profile for this machine called: "{profile.name}"')
+	archinstall.log(f'Found a tailored profile for this machine called: "{_profile.name}"')
 
 	print('Starting install in:')
 	for i in range(10, 0, -1):
@@ -15,4 +15,4 @@ for profile in profile_handler.get_mac_addr_profiles():
 		time.sleep(1)
 
 	install_session = archinstall.storage['installation_session']
-	profile.install(install_session)
+	_profile.install(install_session)
